backend/requirements/base.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.44
psycopg[binary]==3.3.3
alembic==1.15.2
pydantic==2.12.3
pydantic-settings==2.11.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.20
openai==1.76.0
faiss-cpu==1.13.2
numpy==2.3.4
httpx==0.28.1
beautifulsoup4==4.13.4
pypdf==5.4.0
email-validator==2.2.0
redis==5.2.1
dramatiq[redis]==1.17.1
